# docs/swaggerOpenapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Toiage Core — Educational Intelligence Platform",
    "description": "API for generating curriculum-aligned science projects, assessments, and educational content for CBSE classes 6-8.",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/v1/prototypes": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Create prototype",
        "description": "Create a new prototype record.",
        "operationId": "create_prototype_api_v1_prototypes_post",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePrototypeRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "List prototypes",
        "description": "List prototypes with optional filters.",
        "operationId": "list_prototypes_api_v1_prototypes_get",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Status"
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Project Id"
            }
          },
          {
            "name": "assigned_to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Assigned To"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PrototypeResponse"
                  },
                  "title": "Response List Prototypes Api V1 Prototypes Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}": {
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Get prototype by ID",
        "description": "Get a single prototype by ID.",
        "operationId": "get_prototype_api_v1_prototypes__prototype_id__get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/status": {
      "put": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Update prototype status",
        "description": "Update the status of a prototype.",
        "operationId": "update_status_api_v1_prototypes__prototype_id__status_put",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PrototypeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/photos": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Upload prototype photo",
        "description": "Upload a photo for a prototype.",
        "operationId": "upload_photo_api_v1_prototypes__prototype_id__photos_post",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_photo_api_v1_prototypes__prototype_id__photos_post"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "List prototype photos",
        "description": "List all photos for a prototype.",
        "operationId": "list_photos_api_v1_prototypes__prototype_id__photos_get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/notes": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Add prototype note",
        "description": "Add a text note about a prototype build.",
        "operationId": "add_note_api_v1_prototypes__prototype_id__notes_post",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddNoteRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Get prototype notes",
        "description": "Get all notes for a prototype.",
        "operationId": "get_notes_api_v1_prototypes__prototype_id__notes_get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/materials": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Add prototype material",
        "description": "Add a material entry for a prototype.",
        "operationId": "add_material_api_v1_prototypes__prototype_id__materials_post",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMaterialRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Get prototype materials",
        "description": "Get all materials for a prototype.",
        "operationId": "get_materials_api_v1_prototypes__prototype_id__materials_get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/cost-sheets": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Create prototype cost sheet",
        "description": "Create a cost sheet entry for a prototype.",
        "operationId": "create_cost_sheet_api_v1_prototypes__prototype_id__cost_sheets_post",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddCostSheetRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Get prototype cost sheets",
        "description": "Get all cost sheets for a prototype.",
        "operationId": "get_cost_sheets_api_v1_prototypes__prototype_id__cost_sheets_get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/prototypes/{prototype_id}/build-records": {
      "post": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Create prototype build record",
        "description": "Create a build record for a prototype.",
        "operationId": "create_build_record_api_v1_prototypes__prototype_id__build_records_post",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddBuildRecordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "prototype-management"
        ],
        "summary": "Get prototype build records",
        "description": "Get all build records for a prototype.",
        "operationId": "get_build_records_api_v1_prototypes__prototype_id__build_records_get",
        "parameters": [
          {
            "name": "prototype_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Prototype Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {

                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AddMaterialRequest": {
        "properties": {
          "material": {
            "type": "string",
            "title": "Material",
            "description": "Material name",
            "examples": [
              "Cardboard"
            ]
          },
          "quantity": {
            "type": "number",
            "title": "Quantity",
            "description": "Quantity",
            "default": 1
          },
          "unit": {
            "type": "string",
            "title": "Unit",
            "description": "Unit",
            "default": "piece",
            "examples": [
              "sheet",
              "meter",
              "piece"
            ]
          }
        },
        "type": "object",
        "required": [
          "material"
        ],
        "title": "AddMaterialRequest"
      },
      "AddNoteRequest": {
        "properties": {
          "note": {
            "type": "string",
            "title": "Note",
            "description": "Note content",
            "examples": [
              "Cardboard bends under weight"
            ]
          }
        },
        "type": "object",
        "required": [
          "note"
        ],
        "title": "AddNoteRequest"
      },
...
      "Body_upload_photo_api_v1_prototypes__prototype_id__photos_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_photo_api_v1_prototypes__prototype_id__photos_post"
      },
...
      "CreatePrototypeRequest": {
        "properties": {
          "project_id": {
            "type": "string",
            "title": "Project Id",
            "description": "Reference to the science project",
            "default": ""
          },
          "name": {
            "type": "string",
            "title": "Name",
            "description": "Prototype name",
            "examples": [
              "Hydraulic Bridge V1"
            ]
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Description of the prototype",
            "default": ""
          },
          "assigned_to": {
            "type": "string",
            "title": "Assigned To",
            "description": "Who is building this",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "CreatePrototypeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PrototypeResponse": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "project_id": {
            "type": "string",
            "title": "Project Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "assigned_to": {
            "type": "string",
            "title": "Assigned To"
          },
          "total_cost_rs": {
            "type": "number",
            "title": "Total Cost Rs"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "description",
          "status",
          "assigned_to",
          "total_cost_rs",
          "created_at",
          "updated_at"
        ],
        "title": "PrototypeResponse"
      },
...
      "UpdateStatusRequest": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status",
            "description": "New status",
            "examples": [
              "IN_PROGRESS"
            ]
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "UpdateStatusRequest"
      },
...
    }
  }
}